requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7
WTForms==3.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2
requests==2.31.0
email-validator==2.1.0
Pillow==10.0.1
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2

